Simple HUD added

+++ Project/Assets/Scripts/Enemies/Projectile.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Projectile : MonoBehaviour {
 
-	public float dmg = 25;
+	public int dmg = 25;
 
 	// Use this for initialization
 	void Start () {
